Lesson changes
Created a function called setupPuzzle to match all of the data cells in the puzzle and set the initial color of each cell to gold, and to set the cell background color in response to the mousedown event. Called the setupPuzzle function in the init and swapPuzzle functions. Also created the setBackground, extendBackground, and endBackground functions to create and remove an event listener for every puzzle cell.