Lesson changes
Adjusted the setBackground function to prevent the default action selecting a table text, to set the background based on the keyboard event, and to set the cursor type based on the keyboard event. Also adjusted the setupPuzzle function to display a pencil image as the cursor, created object collections of the filled and empty cells, and created an event listener to highlight incorrect cells and display incorrect white cells in pink. Also added an event listener to show the solution button and remove the inline background color style from each cell.